testing more rigourously the neuron
